micro-services/events/requirements.txt
cryptography==44.0.2
firebase-admin==6.6.0
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
gunicorn==23.0.0
PyMySQL==1.1.1
pytest==8.3.5
python-dotenv==1.1.1
flask-cors==5.0.0